more bugfixing
- there is still a small memory leak
- matrix functions get and set are protected against out-of-range
arguments
- weights and biases are initialized with zeros

+++ CNN_seriell/src/Picture.cpp
@@ -32,7 +32,7 @@ Picture::Picture(std::string *line)
 	for(int i=0; i<OUTPUT_SIZE; i++)
 	{
 		std::getline(lineStream,cell, ',');
-		this->output_data[i] = std::stod(cell);
+		this->output_data[i] = (float)std::stod(cell);
 	}
 }
